C programming
Declare 1 integer, 1 float, and 1 doubles variable. Initialize each of the three variables to zero.

Declare two character variables.

Prompt the user to input a value for each variable and store it in that variable.

Output the value of each variable in a statement on one line, left justified and precision of 2.

For example the output should look like this:

The value of the integer is: 234

The value of the float is: 8.93

…..

Notice how the output values are lined up and left justified, you should do the same regardless of the size of the input.